refactor: refactored style-switcher code
