+++ .changeset/warm-panthers-applaud.md
@@ -1,0 +1,5 @@
+---
+"@watergis/svelte-maplibre-style-switcher": patch
+---
+
+refactor: refactored style-switcher code
