+++ packages/style-switcher/src/lib/index.ts
@@ -1,6 +1,4 @@
-import { default as StyleSwitcher } from './StyleSwitcher.svelte';
-import { default as StyleSwitcherControl } from './StyleSwitcherControl.svelte';
-import type { StyleSwitcherOption } from './types';
-import { StyleUrl } from './style-url';
+import StyleSwitcher, { StyleUrl, type StyleSwitcherOption } from './StyleSwitcher.svelte';
+import StyleSwitcherControl from './StyleSwitcherControl.svelte';
 
 export { StyleSwitcher, StyleSwitcherControl, StyleSwitcherOption, StyleUrl };
