--- packages/style-switcher/src/lib/types/Position.ts
@@ -1,1 +1,0 @@
-export type Position = 'top-right' | 'top-left' | 'bottom-right' | 'bottom-left';
