--- packages/style-switcher/src/lib/types/StyleSwitcherOption.ts
@@ -1,4 +1,0 @@
-export type StyleSwitcherOption = {
-	title: string;
-	uri: string;
-};
