--- packages/style-switcher/src/lib/types/index.ts
@@ -1,2 +1,0 @@
-export * from './StyleSwitcherOption';
-export * from './Position';
